+++ classes/ETL/DbModel/Query.php
@@ -20,7 +20,7 @@
  * ==========================================================================================
  */
 
-namespace ETL\DbEntity;
+namespace ETL\DbModel;
 
 use ETL\Utilities;
 use \Log;
