+++ classes/ETL/DbModel/iTableItem.php
@@ -7,7 +7,7 @@
  * ==========================================================================================
  */
 
-namespace ETL\DbEntity;
+namespace ETL\DbModel;
 
 use \Log;
